[gpu][unit_tests]: Refactored PA unit tests.
